Broken status implemented but wonky
Broken elevator seems to not want to drop people

Now people also enter in the elevator's array of people. When it breaks,
it "notifies" the arrival on where the person wants and they get off,
but they're on the wrong floor and they'll repeat their loop.
Now I just need to make it stop waiting for people to ALSO get in.